BASIC (!) support for strings/types/object added.
This commit updates our example 00-print.bas to attempt to store
a string in a variable, then print it.

The evaluator has been largely updated to cope with our new object
primitive, as a quick hack I've disabled FOR/NEXT because they
need to be converted and coding time is full of distractions today!

This is part of #11.