Added complex "pictures".
This commit might require some explaination!  Once upon a time
I bought an ePaper device, only to discover that the library
that drew it used a buffer that had one byte for each pixel
in the display.

That meant that there was no room to :

* Download a PNG/JPG/Image into a buffer.
* Then decode it into the screen area.

Any image would take 2xsize of RAM, which was too much.  So I
decided to cheat:

* Take an image.
* Break that down into a series of lines:
  *  "Draw from 0,0 to 0,10"
  *  "Draw from 0,12 to 0,180"

The saving here was that I could stream and discard this series
of strokes, meaning that RAM was tight but it didn't get exceeded.

Anyway this commit adds the two image-samples I put together to
this project - using 100% identical data.

This closes #54.