Implemented FN.
This is a work-in-progress, because I've only implemented
the calling of a user-defined function - and only then for
a fake one I've created ("square").

That said this works:

        10 PRINT square( 3 ), "\n"

TODO:

* Allow definitions to work.
* Test-cases.
  * Type-check on args?
* More testing.

This is a bit weird because we had to change `factor` to cope
with newlines, due to our child-evaluator.  Makes me wonder ..