Minor tweaks, which break nothing I hope.
I've removed some bounds checking where it was safe, and fixed a single
trailing newline on an error-message thanks to golint.

I think I'm happy that these changes are correct, but the whole point
of this pull-request is that I'll know - the coverage should start to
ramp up again shortly.