Added tests for calling user-defined functions.
OK this commit is weirder than it needs to be, because I've found a
bug.

With these test-cases we see the fake-error:

* End of input while processing term()

Fake?  Well it shouldn't be happening.  In short it is because most
parts of our interpreter look for newline, and the definition of
user-defined functions doesn't have them.  That's swallowed as part
of the parsing process.

Took a while to spot this, but has lead to me removing the implicit
"prefix newline" behavior to the parser, which meant I no longer
skip
the first token in the evaluator - which was always a little odd.

There is a real bug here, and it is to do with the testing of the
end
of term/expr/etc - looking for a newline.  We need to look for
EITHER
newline or EOF.  And this isn't necessarily a bug, or a bounds-fail.

TODO: Handle this properly.