Improve comments and error-checking on this sample
This commit updates the comments within this embedded-example to
better reflect reality; I've also added some more error-checking
to the primitives exported to the BASIC environment.

Specifically we now ensure there are an appropriate number of
arguments before attempting to dereference them - this prevents
fatal errors if a user were to edit the program to read:

    10 CIRCLE 20,30

(i.e. Missing radius value.)