Updated to use the new go1.18+ fuzz-testing
This required making small changes to the fuzz-function, and taking
advantage of the new Context support which was added in #116, to ensure
that the test-cases terminate in a "reasonable" time.

So far I've spotted that the following hangs, and have fixed it:

        PRINT "" * 54555555555555555555

(Issue there is that we run a loop which does the "" + "" + "" ..
step, and that takes "forever" with such a large number.  Given that
the string multiplication of ""*x is always "" we skip that case.)