Updated our string-multiplication.
Fuzzing revealed things like this would take "forever":

        "foo" * 400 * 400 * 400 * 400

Here the naive attempt at skipping the append when the multiplier
was too high failed.  Instead we'll revert to the naive approach
to concatenation, but stop when the string is longer than 64k.

Not ideal to early-terminate, but of course back in the day systems
would only have 64k of RAM (at the most) so strings would never be
that long.